+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @midl-xyz/midl-js-react
 
+## 0.0.7
+
+### Patch Changes
+
+- Add runes hooks
+- Updated dependencies
+  - @midl-xyz/midl-js-core@0.0.5
+
 ## 0.0.6
 
 ### Patch Changes